refactor: domain 패키지로 repository 패키지 이동
