+++ backend/src/main/java/ku_rum/backend/domain/building/service/BuildingSearchService.java
@@ -1,7 +1,7 @@
 package ku_rum.backend.domain.building.service;
 
 import ku_rum.backend.domain.building.dto.response.BuildingResponse;
-import ku_rum.backend.domain.building.repository.BuildingClassRepository;
+import ku_rum.backend.domain.building.domain.repository.BuildingClassRepository;
 import ku_rum.backend.global.exception.building.BuildingNotFoundException;
 import ku_rum.backend.global.exception.building.BuildingNotRegisteredException;
 import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
