+++ backend/src/test/java/ku_rum/backend/domain/friend/application/FriendServiceTest.java
@@ -2,7 +2,7 @@ package ku_rum.backend.domain.friend.application;
 
 import jakarta.transaction.Transactional;
 import ku_rum.backend.domain.building.domain.Building;
-import ku_rum.backend.domain.building.repository.BuildingRepository;
+import ku_rum.backend.domain.building.domain.repository.BuildingRepository;
 import ku_rum.backend.domain.department.domain.Department;
 import ku_rum.backend.domain.department.domain.repository.DepartmentRepository;
 import ku_rum.backend.domain.friend.domain.Friend;
