+++ backend/src/test/java/ku_rum/backend/domain/user/domain/repository/UserRepositoryTest.java
@@ -1,7 +1,7 @@
 package ku_rum.backend.domain.user.domain.repository;
 
 import ku_rum.backend.domain.building.domain.Building;
-import ku_rum.backend.domain.building.repository.BuildingRepository;
+import ku_rum.backend.domain.building.domain.repository.BuildingRepository;
 import ku_rum.backend.domain.department.domain.Department;
 import ku_rum.backend.domain.department.domain.repository.DepartmentRepository;
 import ku_rum.backend.domain.user.domain.User;
